Create custom hooks to move logic around some components and keep them as clean as possible
