+++ src/components/hourly-forecast.tsx
@@ -1,42 +1,14 @@
 'use client';
 
-import { useEffect, useState } from 'react';
-
-import { getHourlyForecast } from '@/actions/get-hourly-forecast';
-import { TempForecast } from '@/lib/definitions/requests';
 import { Subtitle } from './ui/subtitle';
 import { Card } from './ui/card';
-import { useErrorStore, useCityStore } from '@/stores';
 import { formatTemperature } from '@/lib/utils';
+import { useTempForecast } from '@/hooks';
 
 export const HourlyForecast = () => {
-  const { activeCity, units } = useCityStore((state) => state);
-  const [forecast, setForecast] = useState<TempForecast[]>([]);
-  const setError = useErrorStore((state) => state.setError);
+  const { hourlyForecast, activeCity, units } = useTempForecast();
 
-  useEffect(() => {
-    const getClimate = async () => {
-      if (!activeCity) {
-        return;
-      }
-
-      const { lat, lon } = activeCity;
-      try {
-        const res = await getHourlyForecast(lat, lon, units);
-        setForecast(res);
-      } catch (error) {
-        if (error instanceof Error) {
-          setError(error.message);
-        } else {
-          setError('Unknown error');
-        }
-      }
-    };
-
-    getClimate();
-  }, [activeCity, units, setError]);
-
-  if (!forecast.length) {
+  if (!hourlyForecast.length) {
     return null;
   }
 
@@ -54,7 +26,7 @@ export const HourlyForecast = () => {
         aria-label="Hourly temperature forecast scrollable content"
         tabIndex={0}
       >
-        {forecast.map(({ dt_txt, temp }) => {
+        {hourlyForecast.map(({ dt_txt, temp }) => {
           const [date, time] = dt_txt.split('-');
           return (
             <Card
