+++ src/components/search-bar.tsx
@@ -1,72 +1,23 @@
 'use client';
 
-import { useEffect, useState, KeyboardEvent } from 'react';
 import clsx from 'clsx';
 
-import { debounce } from '@/lib/utils';
-import { searchCityLocation } from '@/actions/search-city-location';
-import { City, Units } from '@/lib/definitions/requests';
+import { Units } from '@/lib/definitions/requests';
 import { Subtitle } from './ui/subtitle';
-import { useErrorStore, useCityStore } from '@/stores';
 import { RadioButtons } from './ui/radio-buttons';
+import { useSearchBar } from '@/hooks';
 
 export const SearchBar = () => {
-  const [location, setLocation] = useState<string>('');
-  const [cities, setCities] = useState<City[] | undefined>();
-  const [activeIndex, setActiveIndex] = useState<number>(-1);
-  const setError = useErrorStore((state) => state.setError);
-  const { setActiveCity, setUnits, units } = useCityStore((state) => state);
-  const processLocationChange = debounce((value: string) => setLocation(value));
-
-  const selectActiveCity = (city: City) => {
-    setActiveCity(city);
-    setLocation('');
-    setActiveIndex(-1);
-  };
-
-  const handleKeyDown = (e: KeyboardEvent<HTMLInputElement>) => {
-    if (!cities?.length) return;
-
-    switch (e.key) {
-      case 'ArrowDown':
-        e.preventDefault();
-        setActiveIndex((prev) => (prev < cities.length - 1 ? prev + 1 : prev));
-        break;
-      case 'ArrowUp':
-        e.preventDefault();
-        setActiveIndex((prev) => (prev > 0 ? prev - 1 : prev));
-        break;
-      case 'Enter':
-        e.preventDefault();
-        if (activeIndex >= 0) {
-          selectActiveCity(cities[activeIndex]);
-        }
-        break;
-      case 'Escape':
-        setCities(undefined);
-        setActiveIndex(-1);
-        break;
-    }
-  };
-
-  useEffect(() => {
-    const getCities = async () => {
-      try {
-        const res = await searchCityLocation(location);
-        setCities(res);
-        setActiveIndex(-1);
-      } catch (error: unknown) {
-        const errorMessage = error instanceof Error ? error.message : 'Unknown error';
-        setError(errorMessage);
-      }
-    };
-
-    if (location.length > 0) {
-      getCities();
-    } else {
-      setCities(undefined);
-    }
-  }, [location, setError]);
+  const {
+    location,
+    cities,
+    activeIndex,
+    units,
+    setUnits,
+    selectActiveCity,
+    handleKeyDown,
+    processLocationChange,
+  } = useSearchBar();
 
   return (
     <section
@@ -84,7 +35,13 @@ export const SearchBar = () => {
           aria-controls="city-list"
         >
           <input
-            className="w-full p-2 md:p-3 text-xl rounded md:rounded-lg text-black"
+            className={clsx(
+              'w-full p-2 md:p-3 text-xl rounded md:rounded-lg text-black focus:outline-none focus:ring-2',
+              {
+                'focus:ring-blue-500': cities && cities.length > 0,
+                'focus:ring-red-500': cities?.length === 0,
+              }
+            )}
             type="text"
             placeholder="Enter location"
             onChange={(e) => processLocationChange(e.target.value)}
@@ -94,6 +51,11 @@ export const SearchBar = () => {
             aria-controls="city-list"
             aria-activedescendant={activeIndex >= 0 ? `city-option-${activeIndex}` : undefined}
           />
+          {cities?.length === 0 && (
+            <p role="alert" data-testid="search-error-message" className="p-4 text-red-500">
+              No results found for &quot; {location} &quot;
+            </p>
+          )}
         </div>
         <RadioButtons
           title="Select units of measurement:"
@@ -105,12 +67,6 @@ export const SearchBar = () => {
           changeValue={(value) => setUnits(value as keyof Units)}
           defaultValue={units}
         />
-
-        {cities?.length === 0 && (
-          <p role="alert" data-testid="search-error-message" className="p-4 text-red-500">
-            No results found for &quot; {location} &quot;
-          </p>
-        )}
 
         {cities && (
           <div
