+++ src/components/weather-info.tsx
@@ -1,45 +1,15 @@
 'use client';
 
-import { useEffect, useState } from 'react';
 import Image from 'next/image';
 
-import { getCurrentClimate } from '@/actions/get-current-climate';
-import { Climate } from '@/lib/definitions/requests';
-import { useCityStore, useErrorStore } from '@/stores';
 import { Button } from './ui/button';
-import { formatTemperature, saveActiveCityToLocalStorage } from '@/lib/utils';
+import { formatTemperature } from '@/lib/utils';
+import { useWeatherInfo } from '@/hooks';
 
 const paragraphClasses = 'text-xl mb-3';
 
 export const WeatherInfo = () => {
-  const { activeCity, favCities, units, addFavCities } = useCityStore((state) => state);
-  const setError = useErrorStore((state) => state.setError);
-  const [climate, setClimate] = useState<Climate | undefined>();
-  const addCityToFav = () => {
-    saveActiveCityToLocalStorage(activeCity!);
-    addFavCities([activeCity!]);
-  };
-
-  useEffect(() => {
-    const getClimate = async () => {
-      if (!activeCity) {
-        setClimate(undefined);
-        return;
-      }
-
-      const { lat, lon } = activeCity;
-
-      try {
-        const res = await getCurrentClimate(lat, lon, units);
-        setClimate(res);
-      } catch (error: unknown) {
-        const errorMessage = error instanceof Error ? error.message : 'Unknown error';
-        setError(errorMessage);
-      }
-    };
-
-    getClimate();
-  }, [activeCity, units, setError]);
+  const { climate, favCities, activeCity, units, addCityToFav } = useWeatherInfo();
 
   if (!climate || !activeCity) {
     return null;
